Add files via upload
The PasswordStore protocol is designed to let an owner store and update a private password on-chain and restrict retrieval to the owner. After a manual review of the Smart Contract protocol, I found 2 high level vulnerabilities related to the access management and one informational bug. I learnt that the goal of these security reviews is to improve the protocols security properties but also to teach the companies about security by design.